LSP: slightly improve code
